Document fields that are intentionally missing
To define an FCB sufficient for the Teensy 4, we were able
to skip parts of the FCB. The TODOs describe missing fields that
might be of interest to other users in the future.

I generated this by comparing the FCB layout of the iMXRT1011
to what's implemented. As of this commit, this crate should support
the iMXRT1011 FCB generation. There's no differences from the
iMXRT1062. The iMXRT1062 supports more than the 1011, but those
differences aren't implemented yet.