Migrate to a constant builder approach
This commit removes the API that generates the configuration block
as an array of magic numbers. Instead, it gives users a way to define
configuration blocks at compile time. The entire interface uses const
methods and functions to achieve the compile-time interface.

TODO: fix all the docs, add more tests.