Revise README, API documentation
Feature lists in README and API docs were falling out of sync. Update
the feature list in the API docs, and remove it from the README.

Refer README readers to the API documentation to understand all of this
crate's features. Revise the rationale, and clarify that this crate
doesn't solve the "place in memory" problem.